refactor: extract pg training master
